requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
Django==4.2
django-smtp-ssl==1.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jax==0.4.8
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
msgpack==1.0.5
mysqlclient==2.1.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyotp==2.8.0
python-decouple==3.8
python3-openid==3.2.0
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
zope.interface==6.0
kaggle